pyproject.toml
[project]
name = "playwright-http"
version = "0.1.0"
description = "Browser automation over HTTP - Execute Playwright tests via simple REST API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "playwright>=1.49.0",
    "pytest-playwright>=0.6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["executor"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
